refactor to avoid cyclic import
